requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.55.3,<3.0.0
pydantic==2.5.0
python-dotenv==1.0.0

